requirements.txt
pytest
selenium
requests
allure-pytest
pytest-xdist
webdriver-manager
